+++ swap/src/asb/recovery/cancel.rs
@@ -41,6 +41,8 @@ pub async fn cancel(
 
         // Alice already in final state
         | AliceState::BtcRedeemed
+        | AliceState::XmrRefundTxConstructed { .. }
+        | AliceState::XmrRefundTxPublished { .. }
         | AliceState::XmrRefunded { .. }
         | AliceState::BtcWithholdPublished { .. }
         | AliceState::BtcWithholdConfirmed { .. }