+++ swap/src/asb/recovery/safely_abort.rs
@@ -33,6 +33,8 @@ pub async fn safely_abort(swap_id: Uuid, db: Arc<dyn Database>) -> Result<AliceS
         | AliceState::XmrRefundable { .. }
         | AliceState::BtcPunishable { .. }
         | AliceState::BtcRedeemed
+        | AliceState::XmrRefundTxConstructed { .. }
+        | AliceState::XmrRefundTxPublished { .. }
         | AliceState::XmrRefunded { .. }
         | AliceState::BtcWithholdPublished { .. }
         | AliceState::BtcWithholdConfirmed { .. }