+++ swap/src/cli/cancel_and_refund.rs
@@ -84,6 +84,8 @@ pub async fn cancel(
 
         BobState::Started { .. }
         | BobState::BtcRedeemed(_)
+        | BobState::XmrRedeemConstructed { .. }
+        | BobState::XmrRedeemPublished { .. }
         | BobState::XmrRedeemed { .. }
         | BobState::BtcPunished { .. }
         | BobState::BtcEarlyRefunded { .. }
@@ -225,6 +227,8 @@ pub async fn refund(
         | BobState::SwapSetupCompleted(_)
         | BobState::BtcRedeemed(_)
         | BobState::BtcEarlyRefunded { .. }
+        | BobState::XmrRedeemConstructed { .. }
+        | BobState::XmrRedeemPublished { .. }
         | BobState::XmrRedeemed { .. }
         | BobState::BtcPunished { .. }
         | BobState::BtcWithheld { .. }